Adding filter_molecules.py and updating README documentation
