--- example/README.md
@@ -1,5 +1,0 @@
-# Contents
-# See main README in off-ffcompare for instructions on how to use.
-
-# Daisy will put AlkEthOH stuff in here
-
