--- examples/filter_molecules/README.md
@@ -1,12 +1,0 @@
-# Filtering Large Molecule Sets
-
-We started from three databases of drug like molecules that are available online:
-
-* [DrugBank.sdf](http://www.drugbank.ca/releases/latest)
-* [eMolecules](http://downloads.emolecules.com/free/2016-09-01/)
-* [eMolecules\_incremental](http://downloads.emolecules.com/free/2016-09-01/)
-
-We chose the following requirements for molecules used in this study
-
-* No metal atoms
-* `
